style : feed 페이지 ui 구현
