+++ src/components/common/Title/SubTitle.tsx
@@ -15,6 +15,7 @@ export default function SubTitle({ children, highlightText }: PropsWithChildren<
 const PostingTitleContainer = styled.div`
   display: flex;
   align-items: center;
+
   margin-bottom: 0.5rem;
   ${({ theme }) => theme.fonts.posting_card_title};
